Create Group A Practicle No.6.py
It is decided that weekly greetings are to be furnished to wish the students having their
birthdays in that week. The consolidated sorted list with desired categorical information
is to be provided to the authority. Write a python program to store students PRNs with
date and month of birth. Let List_A and List_B be the two list for two SE Computer
divisions. Lists are sorted on date and month. Merge these two lists into third list
“List_SE_Comp_DOB” resulting in sorted information about Date of Birth of SE Computer
students